requirements.txt
aioredis==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
carehare==1.0.2
certifi==2024.12.14
cffi==1.17.1
channels==3.0.5
channels-rabbitmq==4.0.1
channels-redis==3.3.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==4.2.18
django-bootstrap-v5==1.0.11
django-channels==0.7.0
django-extensions==3.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
h11==0.14.0
hiredis==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pamqp==3.3.0
pika==1.3.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydotplus==2.0.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
zope.interface==7.2
